refactor: resolve #217 - extract useNavigationRoutes composable from GameNavigation
GameNavigation.vue exceeded the 500-line limit (512 lines). Extracted
route grouping logic (NavRoute/GroupedRoutes types, getItemKey, and
groupedRoutes computed) into a dedicated useNavigationRoutes composable.
File reduced from 512 to 459 lines.

Closes #217
